dev/ocr/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# RabbitMQ
aio-pika==9.4.0

# PaddleOCR-VL 1.5 (transformers)
transformers==4.46.0
torch==2.5.1
torchvision==0.20.1
accelerate==1.1.1

# Image processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3

# PDF support
PyMuPDF==1.23.26

# Utils
python-multipart==0.0.6
